Json roundtrip unit test
